changes review - single LoadDirItem class
