+++ full/src/main/java/apoc/load/LoadDirectoryHandler.java
@@ -8,9 +8,25 @@ import org.neo4j.graphdb.Transaction;
 import org.neo4j.kernel.lifecycle.LifecycleAdapter;
 import org.neo4j.logging.Log;
 
-import java.nio.file.*;
-import java.util.*;
-import java.util.concurrent.*;
+import static apoc.load.LoadDirectoryItem.EVENT_KINDS;
+import static apoc.load.LoadDirectoryItem.INTERVAL;
+import static java.nio.file.StandardWatchEventKinds.ENTRY_CREATE;
+import static java.nio.file.StandardWatchEventKinds.ENTRY_DELETE;
+import static java.nio.file.StandardWatchEventKinds.ENTRY_MODIFY;
+import static java.nio.file.WatchEvent.Kind;
+
+import java.nio.file.FileSystems;
+import java.nio.file.Path;
+import java.nio.file.WatchEvent;
+import java.nio.file.WatchKey;
+import java.nio.file.WatchService;
+import java.util.Collections;
+import java.util.HashSet;
+import java.util.List;
+import java.util.Map;
+import java.util.Set;
+import java.util.concurrent.ConcurrentHashMap;
+import java.util.concurrent.Future;
 import java.util.stream.Stream;
 
 import static apoc.util.FileUtils.checkIfUrlEmptyAndGetFileUrl;
@@ -18,7 +34,7 @@ import static apoc.util.FileUtils.getPathFromUrlString;
 
 public class LoadDirectoryHandler extends LifecycleAdapter {
 
-    public final Map<LoadDirectoryItemResult, Future> storage = new ConcurrentHashMap<>();
+    public final Map<LoadDirectoryItem, Future> storage = new ConcurrentHashMap<>();
 
     private final Log log;
     private final GraphDatabaseService db;
@@ -28,6 +44,23 @@ public class LoadDirectoryHandler extends LifecycleAdapter {
         this.db = db;
         this.log = log;
         this.pools = pools;
+    }
+
+    private static Kind[] fromListStringToKindArray(List<String> eventKinds) {
+        Kind[] kinds = eventKinds.stream().map(item -> {
+            switch (item) {
+                case "ENTRY_CREATE":
+                    return ENTRY_CREATE;
+                case "ENTRY_MODIFY":
+                    return ENTRY_MODIFY;
+                case "ENTRY_DELETE":
+                    return ENTRY_DELETE;
+                default:
+                    throw new UnsupportedOperationException("Event Type not supported: " + item);
+            }
+        }).toArray(Kind[]::new);
+
+        return kinds;
     }
 
     @Override
@@ -40,51 +73,50 @@ public class LoadDirectoryHandler extends LifecycleAdapter {
 
     public void remove(String name) {
 
-        final LoadDirectoryItemResult loadDirectoryItemResult = new LoadDirectoryItemResult(name);
-        remove(loadDirectoryItemResult);
+        final LoadDirectoryItem loadDirectoryItem = new LoadDirectoryItem(name);
+        remove(loadDirectoryItem);
     }
 
-    private void remove(LoadDirectoryItemResult loadDirectoryItemResult) {
-        Future removed = storage.remove(loadDirectoryItemResult);
+    private void remove(LoadDirectoryItem loadDirectoryItem) {
+        Future removed = storage.remove(loadDirectoryItem);
         if (removed == null) {
-            String name = loadDirectoryItemResult.getName();
+            String name = loadDirectoryItem.getName();
             throw new RuntimeException("Listener with name: " + name + " doesn't exists");
         }
         removed.cancel(true);
     }
 
-    public void add(LoadDirectoryItemWithConfig loadDirectoryItemWithConfig) {
+    public void add(LoadDirectoryItem loadDirectoryItem) {
 
-        String name = loadDirectoryItemWithConfig.getName();
-        final LoadDirectoryItemResult loadDirectoryItemResult = new LoadDirectoryItemResult(loadDirectoryItemWithConfig);
-
-        if (storage.containsKey(loadDirectoryItemResult)) {
-            remove(name);
+        if (storage.containsKey(loadDirectoryItem)) {
+            remove(loadDirectoryItem);
         }
 
-        storage.put(loadDirectoryItemResult,
-                pools.getDefaultExecutorService().submit(executeListener(loadDirectoryItemWithConfig)));
+        storage.put(loadDirectoryItem, pools.getDefaultExecutorService().submit(executeListener(loadDirectoryItem)));
     }
 
-    public Stream<LoadDirectoryItemResult> list() {
+    public Stream<LoadDirectoryItem> list() {
         return Collections.unmodifiableMap(storage).keySet().stream();
     }
 
     public void removeAll() {
 
-        Set<LoadDirectoryItemResult> keys = new HashSet<>(storage.keySet());
+        Set<LoadDirectoryItem> keys = new HashSet<>(storage.keySet());
         keys.forEach(this::remove);
     }
 
-    private Runnable executeListener(LoadDirectoryItemWithConfig item) {
+    private Runnable executeListener(LoadDirectoryItem item) {
         return () -> {
             try (WatchService watcher = FileSystems.getDefault().newWatchService()) {
 
+                final Map<String, Object> config = item.getConfig();
+                List<String> eventKinds = (List<String>) config.get(EVENT_KINDS);
+                Long interval = (Long) config.get(INTERVAL);
+
                 String urlDir = checkIfUrlEmptyAndGetFileUrl(item.getUrlDir());
-                getPathFromUrlString(urlDir).register(watcher, item.getConfig().getEventKinds());
+                getPathFromUrlString(urlDir).register(watcher, fromListStringToKindArray(eventKinds));
 
-                while (!Thread.currentThread().isInterrupted()) {
-
+                while (true) {
                     WatchKey watchKey = watcher.take();
                     if (watchKey != null) {
                         watchKey.reset();
@@ -111,7 +143,7 @@ public class LoadDirectoryHandler extends LifecycleAdapter {
                             }
                         }
                     }
-                    Thread.sleep(item.getConfig().getInterval());
+                    Thread.sleep(interval);
                 }
             } catch (Exception e) {
                 throw new RuntimeException(e);
