--- full/src/main/java/apoc/load/LoadDirectoryItemWithConfig.java
@@ -1,15 +1,0 @@
-package apoc.load;
-
-public class LoadDirectoryItemWithConfig extends LoadDirectoryItemBase {
-
-    private final LoadDirectoryConfig config;
-
-    public LoadDirectoryItemWithConfig(String name, String pattern, String cypher, String urlDir, LoadDirectoryConfig config) {
-        super(name, pattern, cypher, urlDir);
-        this.config = config;
-    }
-
-    public LoadDirectoryConfig getConfig() {
-        return config;
-    }
-}
